Save and Load Simplification and finishing implementation
-Simplified Save and Load class structure
-Implemented what was left of Play mode Saving and Loading
