+++ Assets/Scripts/Records/AbstractSave.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 28396917b94ebe9428f4b2ce16c986e1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
