+++ Assets/Scripts/Records/Game.cs
@@ -4,14 +4,15 @@ using System.Collections.Generic;
 [Serializable]
 public class Game
 {
+    public string LevelId;
     public PositionCoordinates Player;
     public List<StoryboardStep> Storyboard;
-    //Placeholder have to check how inventory is saved
-    public List<string> Inventory;
+    public List<ItemGroup> Inventory;
 
-    public Game(StoryEngineScript story) {
+    public Game(string level, StoryEngineScript story) {
+        LevelId = level;
         Player = new PositionCoordinates(story.Player.transform.position.x, story.Player.transform.position.z);
         Storyboard = story.Storyboard;
-        Inventory = new List<string>();
+        Inventory = story.StoryItems;
     }
 }
