--- Assets/Scripts/Records/LevelSaveLoad.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0eb6001dcc4e078489e3fb9a03a6d525
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
