+++ Assets/Scripts/Records/PostSaveResponse.cs
@@ -1,0 +1,6 @@
+public class PostSaveResponse {
+    public string id;
+    PostSaveResponse(string id) {
+        this.id = id;
+    }
+}
