+++ Assets/Scripts/Records/PostSaveResponse.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: af2909266f13b5248866f0bf8fb56c83
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
