+++ Assets/Scripts/Records/SaveGame.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8e13e8491c959c54b94654b11f3d0534
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
