+++ Assets/Scripts/Records/SaveLevel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6eb2c9931efe40f47903cd95f87f17c1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
